Image Segments by Letter, Leaves Numbers
Accuracy is much better if I segment the image and read characters
individually. This is harder to do for the numbers since there’s stuff
underneath them too. So right now it will segment out the first four
letters, and then the rest of the image is one big chunk.
Current output: 00L 169665 2’ 226!